changes in localstorage and files
