--- React/game/src/services/loadServices.js
@@ -1,6 +1,0 @@
-const setLocalStorage = (token, userId) => {
-  localStorage.setItem('tokenAuth', token);
-  localStorage.setItem('userId', userId);
-};
-
-export default setLocalStorage;
